requirements.txt
fastapi~=0.115.5
passlib~=1.7.4
pydantic~=2.9.2
PyJWT~=2.9.0
SQLAlchemy~=2.0.36
python-dotenv~=1.0.1
aiohttp~=3.11.6
starlette~=0.41.3
uvicorn~=0.32.0
alembic~=1.14.0